backend/requirements.txt

flask
flask-cors
google-generativeai
sqlite3-binary
